02-airflow/pyproject.toml
[project]
name = "02-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apache-airflow>=3.1.7",
]
